backend/requirements.txt
fastapi
uvicorn
boto3
botocore
pydantic
python-dotenv
pytest
httpx
nova-act
